pyproject.toml
[tool.poetry]
name = "teach-me-concepts-llm"
version = "0.1.0"
description = ""
authors = ["Doug Kim <slakingex@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.4.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
